Separate out position buffer into separate file and document.
